reimplement getUploadUrlByLatestRelease
The octokit getLatestRelease doesn't do what it was expected:
either it filters out draft releases or it seeks for a
release named "latest". Either way, that's not what we want.

So, re-implement it by using listReleases, and getting the
first element, which should be the latest one.

As, for an upload release request we want to see artifacts
even for draft releases, don't filter out the array.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>